// src/lib/v2/types.ts
export type ATSType =
  | "greenhouse"
  | "lever"
  | "ashby"
  | "workable"
  | "teamtailor"
  | "breezy"
  | "smartrecruiters"
  | "bamboohr"
  | "jazzhr";